requirements.txt
alembic
asyncmy
cryptography
fastapi
gunicorn
httpx
itsdangerous
jinja2
passlib
pillow
protobuf==6.31.1
pyotp
python-dotenv
python-jose
python-multipart
qrcode
sqlalchemy[asyncio]
starlette
user-agents
uvicorn[standard]
geoip2
